% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_summary_stats}
\alias{generate_summary_stats}
\title{Generate summary statistics}
\usage{
generate_summary_stats(
  as,
  genotypes,
  facet,
  ac_cols = c("0", "1"),
  fst_cut = 0.95,
  store_pca = FALSE
)
}
\arguments{
\item{as}{allele-count and observed heterozygosity data. Required columns are
\code{variable} containing locus IDs, \code{vf} containing
population/facet IDs, two containing allele counts for
the major/ref and minor/alt loci, respectively, and \code{ho} for observed
heterozygosity. Note that this object can be created from 0/1/2 formatted
genotypic data with \code{prep_as_from_sn}.}

\item{genotypes}{genotypic data, formatted with genotypes as 0/1/2 for the
homozygous ref/major, heterozygous, and homozygous alt/minor, respectively.
Each row must be one locus, each column one individual. Column names must
be unique.}

\item{facet}{character vector noting population informtion for each
inidivudal sample.}

\item{ac_cols}{character, default c("0", "1"). Column names for the two
alleles in 'as' format.}

\item{fst_cut}{numeric, default 0.95. The Fst quantile above which to label
loci as high-Fst.}

\item{store_pca}{logical, default FALSE. If TRUE, PCAs for both all and high-
Fst loci will be retained and returned.}
}
\description{
Generate summary statistics
}
\examples{
#Generate summary statistics
generate_summary_stats(mon_sn, genotypes, facet="population", fst_cut=.95, store_pca = FALSE)
}
\references{
Jombart, T., Devillard, S. & Balloux, F. Discriminant analysis of principal components: a new method for the analysis of genetically structured populations. BMC Genet 11, 94 (2010). https://doi.org/10.1186
}
\author{
William Hemstrom

Andy Lee
}
