% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_p_values}
\alias{get_p_values}
\title{Calculate p-value for the observed PCA}
\usage{
get_p_values(observed, null, h0 = c("less", "greater"))
}
\arguments{
\item{observed}{the observed dataset}

\item{null}{the null dataset}
}
\description{
Calculate p-value for the observed PCA
}
\examples{
# get_p_values(observed, null)

}
\author{
William Hemstrom

Andy Lee
}
