% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{global_fst}
\alias{global_fst}
\title{Calculate global FST}
\usage{
global_fst(x, ac_cols = c("0", "1"))
}
\arguments{
\item{x}{allele-count and observed heterozygosity data. Required columns are
\code{variable} containing locus IDs, \code{vf} containing
population/facet IDs, two containing allele counts for
the major/ref and minor/alt loci, respectively, and \code{ho} for observed
heterozygosity. Note that this object can be created from 0/1/2 formatted
genotypic data with \code{prep_as_from_sn}.}

\item{ac_cols}{character, default c("0", "1"). Column names for the two
alleles in x.}
}
\description{
Calculates FST globally across all subsets
}
\examples{
#global_fst(mon_sn)

}
\author{
William Hemstrom

Andy Lee
}
