% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prep_as_from_sn}
\alias{prep_as_from_sn}
\title{Prepare data in 'as' format from data in 'sn' format}
\usage{
prep_as_from_sn(x, facet)
}
\arguments{
\item{x}{genotypic data, formatted with genotypes as 0/1/2 for the
homozygous ref/major, heterozygous, and homozygous alt/minor, respectively.
Each row must be one locus, each column one individual. Column names must
be unique.}

\item{facet}{character vector noting population inform for each
individual sample.}
}
\description{
Prepare data in 'as' format from data in 'sn' format
}
\examples{
# prep_as_from_sn(mon_sn, facet="populations" )
}
\author{
William Hemstrom

Andy Lee
}
