% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quick_smartPCA}
\alias{quick_smartPCA}
\title{smartPCAs
Generate a raw smartPCA}
\usage{
quick_smartPCA(sn)
}
\arguments{
\item{sn}{allele-count and observed heterozygosity data. Required columns are
\code{variable} containing locus IDs, \code{vf} containing
population/facet IDs, two containing allele counts for
the major/ref and minor/alt loci, respectively, and \code{ho} for observed
heterozygosity. Note that this object can be created from 0/1/2 formatted
genotypic data with \code{prep_as_from_sn}.}
}
\description{
smartPCAs
Generate a raw smartPCA
}
\examples{
# example code, generate a smartPCA
# quick_smartPCA(mon_sn)
}
\author{
William Hemstrom

Andy Lee
}
