% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{run_bootstrapping}
\alias{run_bootstrapping}
\title{Run PCA boostrapping pipeline}
\usage{
run_bootstrapping(x, facet, n, fst_cut = 0.95, par = FALSE, store_pca = FALSE)
}
\arguments{
\item{x}{Genetic data formatted as single numbers indicated genotype, where
0 and 2 are homozygoytes and 1 is a heterozygote. Rows are SNPs and columns
are individuals}
}
\value{
Describe what the function returns to the user.
}
\description{
Runs the full PCA boostrap/permutation testing pipeline using genetic data.
}
\examples{
# provide example code here


}
\references{
# list the citations
}
\author{
William Hemstrom

Andy Lee
}
