% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{run_permutation}
\alias{run_permutation}
\title{Run PCA permutation pipeline}
\usage{
run_permutation(x, facet, n, fst_cut = 0.95, par = FALSE, store_pca = FALSE)
}
\arguments{
\item{x}{Genetic data formatted as single numbers indicated genotype, where 0
and 2 are homozygoytes and 1 is a heterozygote. Rows are SNPs and columns
are individuals.}

\item{facet}{character. Categorical vector by which the data was broken up
e.g., population, treatment. Must contain the same number of samples in the
same order as in \code{x}.}

\item{n}{numeric. Number of permutations to run.}

\item{fst_cut}{numeric, default 0.95. Cut-off used to select highest fst
loci.}

\item{par}{numeric. Number of parallel threads to use.}

\item{store_pca}{logical, default FALSE. If TRUE, returns the raw PCA.}
}
\value{
a list containing: null distribution of the F-statistic observed
  values, the observed F-statistic, the p-values from PCA permutation
  testing, and optionally PCAs for the observed data and all permutations.
}
\description{
Runs the full PCA boostrap/permutation testing pipeline using genetic data.
}
\examples{
run_permutation(snps_dat, "pop", 1000, fst_cut, par = FALSE, store_pca = FALSE)
}
\references{
Patterson, Nick, Price, Alkes L. and Reich, David (2006). Population
Structure and Eigenanalysis. PLOS Genetics 2, 12
\url{https://doi.org/10.1371/journal.pgen.0020190}

Price, Alkes L., Patterson, Nick J., Plenge, Robert M., Weinblatt,
Michael E., Shadick, Nancy A. and Reich, David (2006). Principal components
analysis corrects for stratification in genome-wide association studies.
Nature Genetics 38, 8 \url{https://doi.org/10.1038/ng1847}
}
\author{
William Hemstrom

Andy Lee
}
