% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{mon_sn}
\alias{mon_sn}
\title{monarch butterfly example dataset}
\format{
An object of class \code{data.frame} with 17715 rows and 84 columns.
}
\usage{
mon_sn
}
\description{
A dataset containing genotypes at 17,715 SNP loci for 84 Monarch Butterflies
collected in XXX from several locations across North America,
For more information, see Hemstrom et al (2022).
}
\details{
Stored in the snpRdata object class. See \code{\link{import.snpR.data}} for
details.

Genotypes noted as two characters representing bases sequenced at a specific
site (e.g. "CC" or "GT"). Each SNP occupies one row, each sample occupies on
column. Missing data coded as "NN".
}
\references{
Hemstrom et al (2022). Molecular Ecology. https://doi.org/10.1111/mec.16592
}
\keyword{datasets}
