% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot_permutation_res}
\alias{plot_permutation_res}
\title{Plot PCA permutation results}
\usage{
plot_permutation_res(permute_res, n_boot_pcas = 1)
}
\arguments{
\item{permute_res}{a list containing the null distribution, observed values, and p-vaules from PCA permutation testing}

\item{n_permute_pcas}{number of permutation to plot}
}
\value{
a ggplot of the null distribution of F-statistic and observed F-statistic from PCA permutation testing.
 if do_pcas is \code{TRUE}, also returns a ggplot of the observed PCA and n number of permuted PCAs
}
\description{
Plots results from PCA permutation testing
}
\examples{
# plot observed PCA and 100 permutations
plot_permutation_res(boot_res, 100)

# plot only the observed PCA
plot_permutation_res(boot_res, 0)

}
\author{
William Hemstrom

Andy Lee
}
